Improve mosh-server command line to be more like the "mosh" Perl script.
This adds en_US.UTF-8 as the locale fallback ('-l' option). This should
probably be improved to use the browser user's locale, but this should do for
now.

This also adds the '-s' option, which says to bind to the interface we're
ssh-ing into; this will work better on multihomed systems.

This also adds the '-c 256' option, which brings forth all the beautiful colors
that hterm can display.