Reverting minimum chrome version to 31.
I bumped this to 32 a while back, but it isn't necessary, as I'm not using any
APIs beyond 31. I am using the pepper_33 SDK, but it is to get newer build
tools (e.g., PNaCl exception support), not to get newer APIs. Chrome 31 has UDP
sockets, which is the big special thing we use.

Also, I'm not calling this a new release per se, as this is just a tweak to the
manifest for distribution in the web store. I'm just bumping the build number.