Add "New session" item to launcher menu.
Also close the New Sessions window on connect, instead of minimize. It is no
longer necessary to keep it open, as the user can start a new session either by
launching the app again, or in the case of Chrome OS, right-clicking the app
icon in the "shelf" and choosing "New session".