Add desktop notification for when an update is available.
This may be controversial. If this becomes annoying, I'll rollback this commit,
or provide a way to opt out of notifications. The point here is to help users
know they need to close their Mosh windows to get the latest version, since
users may leave Mosh running for weeks at a time. It is my expectation that
this will only show when Mosh is currently running, otherwise the update should
apply silently. It is also my expectation that this will only show once per new
update. This is rather hard to test, unfortunately.