Major refactor of SSHLogin.
This breaks out the huge SSHLogin function into its own file and class, while
also decomposing it into a set of grokable methods. The interface between
MoshClientInstance and SSHLogin is a bit tortured, and is fodder for a good
rethink and cleanup, but at least the SSHLogin logic is cleaner now.