Update pepper_posix{,_selector}.{cc,h} to avoid raw pointers.
This isn't a total slam dunk. There are still plenty of raw pointers. Some are
there because we're mimicking the unistd API. Some are there as "non-owning raw
pointers", because anything else is rather messy. It may be a sign that the API
needs tweaking, though (e.g, always refer to Targets via IDs, and keep them in
a map).