Remove raw pointers throughout.
This completes (I hope) the removal of all manual memory management, except
where required when working with old APIs (and then properly wrapped). Where
feasible, references were used where ownership is not conveyed. Pointers are
still occasionally used for convenience, but ownership is never conveyed (it is
a bug otherwise).

Of note, this introduces a wrapper for struct msghdr that manages its internal
memory automatically.