Build package with a binary including debug symbols.
This enables connecting to Chrome with gdb for debugging NaCl. Both a
symbol-laden binary and a stripped binary are included in the package, and
Chrome will pick the right one. The symbol-laden binary is not "finalized",
which means it is not in the stable PNaCl bitcode format, so it may only work
with a verion of Chrome which matches the version of the SDK used to build it
(currently 43). But when not running Chrome with --enable-nacl-debug, it will
use the "finalized" version, which is in the stable format.

NB: The debug binary is still built with flags that upset the debugger. During
development, consider temporarilty tweaking the flags (see the Makefile for
details).

Also, this commit includes a convenience GDB script which facilitates
connecting gdb to a local Chrome.