Build pre-translated .nexe files.
This avoids the overhead of the client doing this translation at first runtime.

Moreover, this is necessary because the pnacl-translate tool for ARM, which
usually runs on the client the first time it sees the .pexe, builds broken
.nexes sometimes. This may explain the crashes seen in Issue #90. Building an
already-translated .nexe eliminates variances among installations.

When run with -threads=seq, it builds .nexes that at least pass the "smoke
test" 6/10 times. It seemed to fail more often in default mode (which uses more
threads while translating), but I didn't test thoroghly. But one interesting
observation is that it builds .nexes of different sizes unless run with
-threads=seq. Notably, when building for other architectures, the .nexes are
always the same size in default mode. Unfortunately, regardless of architecture
or threads setting, the .nexes are not built deterministically.