Initial SSHFP implementation.
This parses the SSHFP data and can validate against a key. This is not wired
in yet, though.

This also introduces the first unit test. \o/ But it isn't wired in. :'(
Bazel doesn't quite know how to build and run a test locally due to the whole
cross-compiling thing. It is totally within Bazel's core ability, but I need to
spend a little time on it. For now, I just built and ran it by hand.