Provide migration path for fingerprints stored by IP addr.
Fingerprints are now (more properly) stored by hostname. But it will surprise
users if suddenly Mosh for Chrome appears to forget all their fingerprints. So
this provides a migration path.

I think it would be bad form to import these fingerprints without user
intervention, so there is a prompt letting the user know what is going on.