Move Bazel flags into tools/bazel.rc, and clean up build.sh.
It is now reasonable to build directly with the bazel command; e.g.:

  $ bazel build :mosh_chrome_dev

For doing a build that is going to be distrubited, it is best to continue to
use build.sh, though, just to ensure it is built properly.

The benefit of running bazel directly is that you can do quicker "sanity-check"
builds easily. E.g.:

  $ bazel build mosh_nacl:ssh_lib

This avoids having to link and translate the whole binary, just to see that the
bit you're working on compiles.