Add BUILD rule for unit test for sshfp_record.
All tests can be run like so:
  $ bazel test mosh_nacl:sshfp_record_test --cpu=k8

I wish you didn't have to specify the CPU. Also, if you want to test the whole
tree (which only has this one test atm), you have to add --build_tests_only,
otherwise it'll try to build targes which aren't appropriate for k8 (e.g.,
//mosh_nacl:mosh_chrome). Not sure why Bazel would bother building targets
which aren't needed for tests.

Had to undefine NDEBUG for unit test so that assert() woudl work. The test
really shouldn't be using assert(), but it is. And tests ordinarily aren't
built in opt mode (which defines NDEBUG, disabling assert()), but it is by
default due to //tools/bazel.rc.