Cause app to update itself promptly once no sessions are active.
It can take a long while for the background page to unload, and until it does,
the app will not update. So that annoying message saying to close the app to
get the update sticks around for a long time, and the user is unlikely to know
how to get rid of it. This is especially bad since mosh sessions tend to
persist for a long while.

This commit causes the app to update itself immediately if there are no
sessions running, or immediately upon the close of the last running session.

Fixes #93.