Fix mouse events for row/column values >= 96.
Strictly speaking, 96 is the proper limit, as I believe terminal escape
sequences all need to be 7-bit ASCII, and due to the way the row/column value
is encoded, the 8th bit is invoked at or past 96. But vim (and possibly others)
accept higher values.

When conveying input data between hterm and NaCl, the string became UTF-8-ified.
This broke the escape encoding for high-order codepoints. Altered the manner in
which this data is conveyed to avoid such cleverness.

Fixes #131.