Make work with latest released Bazel 0.14.1.
Versions after 0.3.1 did strict dependency enforcement of includes. But the
compiler in the NaCl SDK doesn't get along with it, so it is effectively (and
hackishly) disabled in the CROSSTOOL file. Given the state of NaCl, I don't
think it'll ever be fixed.

As a consequence of doing this, tool out the io_bazel repo, which was used for
protobufs and gtest, and replace it with those two projects directly. Those
projects now have their own BUILD files.

Had to eliminate -Wunused-parameter to get protobufs to compile. Oh well.

Fixes #140.