fix minor crash with term-init-string prefs
The storage call here will return an empty object when the key
isn't found which causes this code to pass undefined down to the
terminal print layers (which then crashes).  It doesn't affect
the running of the app because this is a callback func in its
own execution scope, but it does clutter up the console log.