Move to Bazel 0.23.0.
This is the last version of Bazel to support the use of the CROSSTOOL file. The
next upgrade migth be a doozy.

It was necessary to bump the version of protobuf in order to get past some
deprecated usage:
https://github.com/bazelbuild/bazel/issues/5827

This further necessitated adding some transitive dependencies. (This is solved
more properly in a later version of protobuf, but moving that far came with a
number of other challenges.) This also required setting
`-Wno-inconsistent-missing-override`; I hope that warning can be restored in
the future.