Bump Bazel all the way to 5.0.0.
This brings the project up-to-date with the latest Bazel version.

No changes were required going straight from 1.0.0 to 5.0.0! The only change
here is a cleanup removal of `gcov` from the toolchain definition, as it is was
stubbed out anyway and is optional as of Bazel 5.0.0.

Also tested fine with `bazelisk --strict`.